Add validation

+++ QuestionSet/Products/v1/IncomeProtectionV1.cs
@@ -16,7 +16,7 @@
 
         public bool HasBeenReferredToPsychiatrist { get; set; }
 
-        public int HeightInMeter { get; set; }
+        public decimal HeightInMeter { get; set; }
 
         public decimal WeightInKg { get; set; }
